django/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2024.7.4
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-axes==8.0.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-prometheus==2.4.0
django-recaptcha==4.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.12.3
idna==3.10
kombu==5.5.3
pillow==11.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==8.2
redis==5.0.7
requests==2.32.3
sentry-sdk[django]==2.10.0
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
